Add custom build functionality
