+++ .eslintignore
@@ -1,3 +1,2 @@
 node_modules/**/*
 dist
-webpack.config.js
