+++ README.md
@@ -11,6 +11,9 @@ easily be used from Webpack, RequireJS, Browserify, Node etc. Since it is
 written using ES2015+, you can also import the ES module directly from your own
 ES2015+ project.
 
+The included bundle has all functionality built in but you can easiliy make a
+custom build that suits your project's needs. See below for instructions.
+
 **Notes for exif-js users**
 
 If you come here from the popular but now dead exif-js package, please let me
@@ -139,6 +142,65 @@ fs.writeFileSync('/path/to/new/thumbnail.jpg', Buffer.from(tags['Thumbnail'].ima
 
 See [examples/](examples/) directory for more details.
 
+### Configure a custom build
+
+This is for npm users. To specify what functionality you want you can either use
+include pattern (start with an empty set and include) or exclude pattern (start
+with full functionality and exclude). If an include pattern is set, excludes
+will not be used.
+
+**Example 1:** Only include JPEG files and Exif tags (this makes the bundle almost
+half the size of the full one (non-gzipped)):
+
+```bash
+npm config set exifreader:include jpeg,exif
+```
+
+**Example 2:** Exclude XMP tags:
+
+```bash
+npm config set exifreader:exclude xmp
+```
+
+**Example 3:** Delete a setting:
+
+```bash
+npm config delete exifreader:include
+```
+
+Then, if you didn't install ExifReader yet, just run the `npm install` command
+from above. Otherwise you have to re-build the library:
+
+```bash
+npm rebuild exifreader
+```
+
+After that the new bundle is here: `node_modules/exifreader/dist/exif-reader.js`
+
+**NOTE:** The configuration will be saved in your user's home directory
+(in the file `.npmrc`), not in the project's directory, so make sure you have something set
+up that reproduces this when you build for production.
+
+If you're using the include pattern config, remember to include everything you
+want to use. If you want `xmp` and don't specify any file types, you will get
+"Invalid image format", and if you specify `jpeg` but don't mention any tag
+types no tags will be found.
+
+Possible modules to include or exclude:
+
+| Module      | Description                   |
+| ----------- | ----------------------------- |
+| `jpeg`      | JPEG images.                  |
+| `tiff`      | TIFF images.                  |
+| `png`       | PNG images.                   |
+| `heic`      | HEIC/HEIF images.             |
+| `file`      | JPEG image width, height etc. |
+| `exif`      | Regular Exif tags.            |
+| `iptc`      | IPTC tags.                    |
+| `xmp`       | XMP tags.                     |
+| `icc`       | ICC color profile tags.       |
+| `thumbnail` | JPEG thumbnail.               |
+
 Notes
 -----
 
@@ -217,6 +279,8 @@ npm run coverage
 Known Limitations
 -----------------
 
+-   For PNG files, only uncompressed XMP tags are currently supported.
+-   For HEIC files, only Exif tags are currently supported.
 -   The descriptions for UserComment, GPSProcessingMethod and GPSAreaInformation
     are missing for other encodings than ASCII.
 
@@ -247,6 +311,7 @@ Changelog
 ---------
 
 -   **March 2020**:
+    -   Add functionality to create a custom build to reduce bundle size.
     -   Add support for PNG images.
     -   Major update to version 3.0. However, the actual change is quite small,
         albeit a breaking one if you use that functionality (`.value` on
