+++ package.json
@@ -32,6 +32,7 @@
     "mocha": "^7.1.1",
     "npm-run-all": "^4.1.5",
     "nyc": "^14.1.1",
+    "string-replace-loader": "^2.2.0",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.6"
   },
@@ -40,7 +41,8 @@
     "coverage": "nyc npm test",
     "lint": "eslint .",
     "precommit": "npm-run-all lint test",
-    "test": "cross-env BABEL_ENV=test mocha --require @babel/register"
+    "test": "cross-env BABEL_ENV=test mocha --require @babel/register",
+    "postinstall": "npm run build"
   },
   "nyc": {
     "check-coverage": true,
