+++ src/constants.js
@@ -1,0 +1,11 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+export default {
+    USE_FILE: true,
+    USE_EXIF: true,
+    USE_IPTC: true,
+    USE_XMP: true,
+    USE_ICC: true
+};
